requirements-dev.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
boto3==1.5.22
botocore==1.8.36
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
dj-static==0.0.6
Django==2.0.2
django-storages==1.6.5
docutils==0.14
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
jmespath==0.9.3
lxml==4.1.1
parsel==1.3.1
Pillow==5.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
queuelib==1.4.2
s3transfer==0.1.12
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
static3==0.7.0
Twisted==17.9.0
unicodecsv==0.14.1
urllib3==1.22
w3lib==1.19.0
zope.interface==4.4.3
